Created searchbar component for Navbar. Additional navbar width when on med - mobile sizes causes content blockage - commenting out component until resolved.
