+++ client/src/pages/Navbar/Navbar.jsx
@@ -1,13 +1,16 @@
-import { Typography, AppBar, Box, Toolbar, Container } from '@mui/material';
+import { Typography, AppBar, Box, Toolbar, Container, useMediaQuery, useTheme } from '@mui/material';
 import { NavLink } from 'react-router-dom';
 import LogoDevIcon from '@mui/icons-material/LogoDev'; // Logo placeholder
 import MenuDrawer from './Navigation/MenuDrawer';
 import TopNavbar from './Navigation/TopNavbar';
 import CartDrawer from './CartDrawer';
 import { useAuthContext } from '../../hooks/useAuthContext';
+import SearchBar from './Search/Search'; // Custom SearchBar component
 
 export default function Navbar() {
   const { user } = useAuthContext();
+  const theme = useTheme();
+  const isMobile = useMediaQuery(theme.breakpoints.down('md')); // mediaQuery hook for mobile/md size 
 
   return (
     <Box sx={{ display: 'flex' }}>
@@ -17,42 +20,54 @@ export default function Navbar() {
         elevation={0}
       >
         <Container maxWidth='xl'>
-          <Toolbar disableGutters>
-            {/* Menu drawer (mobile) */}
-            <MenuDrawer />
+    
+            <Toolbar>
 
-            {/* Logo & Brand Name */}
-            <Typography
-              variant='h6'
-              component='div'
-              sx={{
-                flexGrow: 1,
-                textAlign: { xs: 'center', sm: 'left' },
-              }}
-            >
-              {user ? (
-                <NavLink
-                  to='/explore/all'
-                  style={{ textDecoration: 'none', color: '#fff' }}
-                >
-                  <LogoDevIcon /> AppName
-                </NavLink>
-              ) : (
-                <NavLink
-                  to='/'
-                  style={{ textDecoration: 'none', color: '#fff' }}
-                >
-                  <LogoDevIcon /> AppName
-                </NavLink>
-              )}
-            </Typography>
+              {/* Menu drawer */}
+              <MenuDrawer />
 
-            {/* Main navbar */}
-            <TopNavbar />
+              {/* Logo & Brand Name */}
+              <Typography
+                variant='h6'
+                component='div'
+                sx={{
+                  flexGrow: 1,
+                  textAlign: { xs: 'center', sm: 'left' },
+                }}
+              >
+                {user ? (
+                  <NavLink
+                    to='/explore/all'
+                    style={{ textDecoration: 'none', color: '#fff' }}
+                  >
+                    <LogoDevIcon />E-Commerce
+                  </NavLink> 
+                ) : (
+                  <NavLink
+                    to='/'
+                    style={{ textDecoration: 'none', color: '#fff' }}
+                  >
+                    <LogoDevIcon />E-Commerce
+                  </NavLink>
+                )}
+              </Typography>
 
-            {/* Checkout drawer */}
-            <CartDrawer />
-          </Toolbar>
+              {/* SearchBar for non-mobile screens */}
+              {/* {!isMobile && <SearchBar/>} */}
+
+              {/* Main navbar */}
+              <TopNavbar />
+
+              {/* Checkout drawer */}
+              <CartDrawer />
+
+
+            </Toolbar>
+
+            {/* SearchBar for mobile screens */}
+            {/* {isMobile && <SearchBar/>} */}
+
+         
         </Container>
       </AppBar>
     </Box>
